+++ release-notes.md
@@ -8,6 +8,9 @@ Changes:
   - New configuration "UiOptions.AutoAddWildcardsToSearchFields" (default = true)
     to enable/disable automatic adding of wildcards before and after the value entered in the UI search fields
     for textual values.
+  - New configurations to customize the filenames of the downloads:
+    "DownloadStudyFileNameTemplate", "DownloadSeriesFileNameTemplate" and "DownloadInstanceFileNameTemplate".
+    The last configuration will only work with an Orthanc 1.12.7+.
   - New configuration "Tokens.RequiredForLinks" can be set to false
     when using HTTP Basic authentication together with the authorization plugin
     (https://discourse.orthanc-server.org/t/user-based-access-control-with-label-based-resource-access/5454).